Small hygiene update.
